codebert-base.commons-lang.mid-3439.idx-21201.2.mutant

+++ commons-lang/src/main/java/org/apache/commons/lang3/time/StopWatch.java
@@ -327,7 +327,7 @@ public class StopWatch {
      * @since 2.1
      */
     public long getSplitTime() {
-        return getSplitNanoTime() / NANO_2_MILLIS;
+        return (int) (getSplitNanoTime() / NANO_2_MILLIS);
     }
 
     /**
